op-node: Move withdrawal utils out of node into op-e2e
These are intended to be utilities for e2e tests so shouldn't live with the production code in op-node.
