+++ op-e2e/actions/user.go
@@ -8,6 +8,7 @@ import (
 	"math/rand"
 	"time"
 
+	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils/withdrawals"
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -26,7 +27,6 @@ import (
 	"github.com/ethereum-optimism/optimism/op-e2e/config"
 	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/derive"
-	"github.com/ethereum-optimism/optimism/op-node/withdrawals"
 )
 
 type L1Bindings struct {
